Dash_mosarch/requirements.txt
gunicorn==21.2.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
dash==2.13.0
dash_extensions==1.0.4
pandas==1.5.3
numpy==1.24.3
pillow==10.2.0
onnxruntime==1.17.1


